extract allies
